+++ resource/commands/create-buildpack.sh
@@ -7,7 +7,6 @@ enabled=$(get_option '.enabled')
 logger::info "Executing $(logger::highlight "$command"): $buildpack"
 
 args=("$buildpack" $path "$position") # don't quote $path so we can support globbing
-[ "$enabled" == "true" ]  && args+=(--enable)
 [ "$enabled" == "false" ] && args+=(--disable)
 
 cf::cf create-buildpack "${args[@]}"