+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
-- `run-task`, `add-network-policy`, `remove-network-policy`, `create-route` commands now work using `cf_cli_version: 7`
+- `run-task`, `add-network-policy`, `remove-network-policy`, `create-route`, `create-buildpack` commands now work using `cf_cli_version: 7`
 
 ### Packaged Dependencies